\encoding{UTF-8}
\name{bournemouth}
\alias{bournemouth}
\title{Air Pollution Measurements in Bournemouth}
\description{
The \code{bournemouth} data frame contains daily air pollution
measurements in Bournemouth UK from 2004 to 2023.
}
\format{
  A data frame with 1805 observations on 3 variables:
  \describe{
    \item{\code{Date}}{Calendar date in YYYYMMDD format.}
    \item{\code{Nitric.oxide}}{Daily concentration of nitric oxide (NO).}
    \item{\code{Nitrogen.dioxide}}{Daily concentration of nitrogen dioxide (NO\eqn{_2}).}
  }
}
\details{
  Use \code{dataset("bournemouth")} to load these data from GitHub. 
}
\references{
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2026).
  \emph{Handbook of Statistics of Extremes}. Chapman & Hall/CRC, Boca
  Raton, FL.

  Simpson, E. S., and Wadsworth, J. L. (2026).
  Conditional extremes modeling. In: \emph{Handbook of Statistics of Extremes},
  Chapter 10, pp. 199--220.
}
\keyword{Environmental Sciences}
