\encoding{UTF-8}
\name{brainwave}
\alias{brainwave}
\title{Brainwave data}
\description{
  The data contains the EEG power of two commonly-recognized types of
  EEG frequency bands: \code{Y1} for alpha and \code{Y2} for beta, for 30 participants and different covariates/stimulus. Column 3 to 8, represent the stimulus in the set: \code{x1} for "mathematics", \code{x2} for "relaxation", \code{x3} for "music", \code{x4} for "color", \code{x5} for "video", \code{x6} for "think and relax"). Column 9 represents the \code{id} of the participant, and column 10 contains the \code{time} in seconds.
}
\usage{brainwave}
\format{
  The \code{brainwave} data frame has 7506 rows and 10 columns.
}
\references{
  Palacios Ramirez, V., de Carvalho, M., and Gutierrez, L. (2024, to appear)
  Heavy-tailed NGG-mixture models. \emph{Bayesian Analysis}. 
}
%% \source{
%%  John Chuang, Nick Merrill, Thomas Maillart, and Students of the UC Berkeley Spring 2015 MIDS Immersion Class. "Synchronized Brainwave Recordings from a Group Presented with a Common Audio-Visual Stimulus (May 9, 2015)." May 2015. }
\keyword{Medicine}
