\name{california}
\alias{california}
\title{California Fire Perimeters}
\description{
  The \code{california} data frame has 16577 rows and 2 columns.
  The first column contains the date, the second
  column gives the quantity of acres consumed by the flames.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Date}{A numeric vector of dates of wildfires.}
    \item{Acres}{A numeric vector of thousands of acres consumed by the
      flames.}
  }
}
\details{
  Data are available from GitHub
  and hence can be gathered using the command \code{dataset("california")}. 
}
\references{
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2026).
  \emph{Handbook on Statistics of Extremes}. Chapman & Hall/CRC. Boca Raton, FL.

  de Carvalho, M., Palacios, V., Henriques-Rodrigues, L., and Lee, M. W. (2026).
  Regression Models for Extreme Events.
  In: Handbook of Statistics of Extremes, Chapter 6, pp. 99--120.
}
\source{
  California State Geoportal.
}
\keyword{Forestry}
