\encoding{UTF-8}
\name{DATAstudio-package}
\alias{DATAstudio-package}
\alias{DATAstudio}
\docType{package}
\title{The Research Data Warehouse of Miguel de Carvalho}
\description{
  \if{html}{
    \figure{logo.png}{options: alt='logo' width='120'}
  }  
  \if{latex}{\figure{logo.png}{options: width=3cm}}
  
  \pkg{DATAstudio} is an add-on tool for \R that pulls together 
  a collection of datasets used in Miguel de Carvalho's research. 
  For a complete list of datasets and documentation, type \code{help.start()} 
  and follow the link to \pkg{DATAstudio} on the Package Index.

  If you use data from this package in publications,
  please cite the package and the references provided below.
}
\author{Miguel de Carvalho; School of Mathematics, University of Edinburgh.}
\seealso{
  \url{https://www.maths.ed.ac.uk/~mdecarv/}
}
\section{Funding}{
  Generative AI Lab (Univ. of Edinburgh). Royal Society of Edinburgh. 
}
\keyword{DATAstudio}
