\encoding{UTF-8}
\name{hongkong}
\alias{hongkong}
\title{Daily Maximum Temperature in Hong Kong}
\description{   
  Daily Maximum Temperature Data from Hong Kong International Airport,
  Hong Kong, from January 1884 to October 2023.
}
\format{
  The \code{hongkong} data frame has 48517 observations and 2 columns:
  \describe{
    \item{\code{date}}{
      Year-month-day.
    }
    \item{\code{value}}{  
      Daily maximum temperature (in degrees Celsius). 
    }
  }
}
\details{
  Data on daily maximum temperatures with no missing values,
  with a total of 48517 observations. Data are available from GitHub
  and hence can be gathered using the command \code{dataset("hongkong")}. 
}
\references{
  Carcaiso, V., De Carvalho, M., Prosdocimi, I. and Antoniano-Villalobos,
  I. (2026). Bayesian mixture models for heterogeneous extremes. arXiv:2509.15359.
  
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2026).
  \emph{Handbook on Statistics of Extremes}. Chapman &
  Hall/CRC. Boca Raton, FL.
}
\keyword{Climatology & Meteorology}
