\encoding{UTF-8}
\name{kfrench}
\alias{kfrench}
\title{Fama--French Industry Portfolio Returns}
\description{
  The \code{kfrench} data frame contains daily returns for 30
  Fama--French industry portfolios from Jan 1970 to December 2023.
}
\format{
  A data frame with 13 599 observations on 31 variables:
\describe{
  \item{\code{time}}{Trading day in YYYYMMDD format.}
  \item{Industry portfolios}{Thirty daily Fama--French industry portfolio
    return series, in percent.}
  }
}
\details{
  Use \code{dataset("kfrench")} to load these data from GitHub.
}
\references{
  Cooley, D., Sabourin, A., and Wixson, T. (2026).
  Principal component analysis for multivariate extremes. In:
  \emph{Handbook of Statistics of Extremes}, Chapter 11, pp. 221--242.

  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2026).
  \emph{Handbook of Statistics of Extremes}. Chapman & Hall/CRC, Boca
  Raton, FL.
}
\keyword{Finance}
