\encoding{UTF-8}
\name{lse}
\alias{lse}
\title{Selected Stocks from the London Stock Exchange}
\description{
  Prices at close from 26 selected stocks from the London stock exchange
  from 1989 to 2016. 
}
\usage{lse}
\format{
  The \code{lse} data frame has 6894 rows and 27 columns.
}
\references{
  Rubio, R., de Carvalho, M., and Huser (2018) Similarity-based
  clustering of extreme losses from the London stock exchange.
}
\keyword{Financial}
