\encoding{UTF-8}
\name{passengers}
\alias{passengers}
\title{International Airline Traffic Data}
\description{
  Monthly number of passengers (in thousands) in a group of several
  international airline companies from January 1949-December 1960.
}
\usage{passengers}
\format{
  A time series with 144 observations; the object is of class \code{ts}.
}
\references{
  Brown, R.G. (1963) \emph{Smoothing, Forecasting and Prediction of
    Discrete Time Series}. New Jersey: Prentice-Hall.

  Rodrigues, P. C. and de Carvalho, M. (2013) Spectral modeling of time
  series with missing data. \emph{Applied Mathematical Modelling},
  \bold{37}, 4676-4684.
}
\keyword{Business}
