\encoding{UTF-8}
\name{santiago}
\alias{santiago}
\title{Santiago Temperature Data}
\description{
  The data consist of average daily air temperatures in Fahrenheit
  scale---rounded to the nearest integer---of Santiago (Chile) from
  April 1990 to March 2017.
}
\usage{santiago}
\format{
  A dataframe with 10126 observations on one variable.
}
\references{
  Galasso, B., Zemel, Y. and de Carvalho, M. (2022). Bayesian
  semiparametric modelling of phase-varying point processes.
  \emph{Electronic Journal of Statistics}, \bold{16}, 2518-2549.
}
\source{
  NOAA's National Centers for Environmental Information (NCEI).
}
\keyword{Climatological}
