\encoding{UTF-8}
\name{sp500}
\alias{sp500}
\title{Standard & Poor 500}
\description{
  Daily Standard and Poor’s index at close from 1988 till 2007. 
}
\usage{sp500}
\format{
  The \code{sp500} data frame has 5043 rows and 2 columns: \code{date}
  and price at \code{close}.
}
\references{
  de Carvalho, M. (2016) Statistics of extremes: Challenges and
  opportunities.  In: \emph{Handbook of EVT and its Applications to Finance
    and Insurance}. Eds F. Longin. Hoboken: Wiley.
}
\keyword{Financial}
