\encoding{UTF-8}
\name{sp500a}
\alias{sp500a}
\title{Standard & Poor's 500 and Sector Indices}
\description{
Daily S&P 500 index at close from 2002 till 2024 along with sector indices.
}
\format{
A data frame with 5552 observations on 13 variables:
\describe{
  \item{\code{Date}}{Trading day.}
  \item{\code{S.P.Market}}{S&P 500 index level.}
  \item{\code{S.P.Communication.Services}}{Communication Services sector index.}
  \item{\code{S.P.Technology}}{Information Technology sector index.}
  \item{\code{S.P.Industrial}}{Industrials sector index.}
  \item{\code{S.P.Materials}}{Materials sector index.}
  \item{\code{S.P.Consumer.Discretionary}}{Consumer Discretionary sector index.}
  \item{\code{S.P.Financial}}{Financials sector index.}
  \item{\code{S.P.Health.Care}}{Health Care sector index.}
  \item{\code{S.P.Consumer.Staples}}{Consumer Staples sector index.}
  \item{\code{S.P.Utilities}}{Utilities sector index.}
  \item{\code{S.P.Real.Estate}}{Real Estate sector index.}
  \item{\code{S.P.Energy}}{Energy sector index.}
}
}
\details{
  Use \code{dataset("sp500a")} to load these data from GitHub.
}
\references{
  Davison, A., and Miralles, O. (2026).  Modeling univariate
  extremes---why and how. In: \emph{Handbook of Statistics of Extremes},
  Chapter 2, pp. 11--35.

  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2026).
  \emph{Handbook of Statistics of Extremes}. Chapman & Hall/CRC, Boca
  Raton, FL.
}
\keyword{Finance}
