\encoding{UTF-8}
\name{tmt}
\alias{tmt}
\title{Trail Making Test}
\description{
  Completion times in seconds for TMT (Trail Making Test), part A, for 245 
  patients with Parkinson's disease, along with corresponding diagnostic on
  cognitive impairment.
}
\usage{tmt}
\format{
  The \code{tmt} data frame has 245 rows and 2 columns: 	
  \describe{
    \item{\code{marker}}{
      completion times (in seconds)
    }
    \item{\code{status}}{
      disease status of each subject, with \code{1}, \code{2}, and
      \code{3} respectively denoting patients diagnosed as unimpaired,
      mild cognitive impairment, and dementia.
    }
  }
}
\references{
  Inácio de Carvalho, V., de Carvalho, M., and Branscum, A. (2018) 
  Bayesian bootstrap inference for the ROC surface.
  \emph{Stat}, \bold{7}, e211.
}
\keyword{Medical}
