\name{california}
\alias{california}
\title{California Fire Perimeters}
\usage{data(california)}
\description{
  The \code{california} data frame has 16577 rows and 2 columns.
  The first column contains the date, the second
  column gives the quantity of acres consumed by the flames.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Date}{A numeric vector of dates of wildfires.}
    \item{Acres}{A numeric vector of thousands of acres consumed by the
      flames.}
  }
}
\references{
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2025, to
  appear). \emph{Handbook on Statistics of Extremes}. Chapman &
  Hall/CRC. Boca Raton, FL.
}
\source{
  California State Geoportal.
}
\keyword{Forestry}
