\encoding{UTF-8}
\name{DATAstudio-package}
\alias{DATAstudio-package}
\alias{DATAstudio}
\docType{package}
\title{The Research Data Warehouse of Miguel de Carvalho}
\description{
  \if{html}{
    \figure{logo.png}{options: alt='logo' width='120'}
  }  
  \if{latex}{\figure{logo.png}{options: width=3cm}}
  
  \pkg{DATAstudio} is an add-on tool for \R that pulls together 
  a collection of datasets used in Miguel de Carvalho's research. 
  For a complete list of datasets and documentation, type \code{help.start()} 
  and follow the link to \pkg{DATAstudio} on the Package Index. 
}
\author{Miguel de Carvalho; School of Mathematics, University of Edinburgh.}
\seealso{
  Useful links:
  \itemize{
    \item \url{https://www.maths.ed.ac.uk/~mdecarv/publications.html}
  }
}
\section{Funding}{
  Fundação para a Ciência e a Tecnologia (Portuguese NSF) grants:
  \itemize{
    \item PTDC/MAT-STA/28649/2017.
    \item UID/MAT/00006/2019.
  }
}
\keyword{DATAstudio}
