\encoding{UTF-8}
\name{hongkong}
\alias{hongkong}
\title{Daily Maximum Temperature in Hong Kong}
\description{   
  Daily Maximum Temperature Data from Hong Kong International Airport,
  Hong Kong, from January 1884 to October 2023.
}
\usage{hongkong}
\format{
  The \code{hongkong} data frame has 48517 observations and 2 columns:
  \describe{
    \item{\code{date}}{: 
      Year-month-day.
    }
    \item{\code{value}}{: 
      Daily maximum temperature (in degrees Celsius). 
    }
  }
}
\details{
  Data on daily maximum temperatures with no missing values,
  with a total of 48517 observations. 
}
\references{
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2025, to
  appear). \emph{Handbook on Statistics of Extremes}. Chapman &
  Hall/CRC. Boca Raton, FL.
}
\keyword{Climatology & Meteorology}
