\encoding{UTF-8}
\name{lisbon}
\alias{lisbon}
\title{Rainfall Data from Lisbon, Portugal}
\description{   
  Daily rainfall data from Lisbon, Portugal, from December 1863 to June 2018.
}
\usage{lisbon}
\format{
  The \code{lisbon} data frame has 56503 observations and 2 columns:
  \describe{
    \item{\code{yearmonth}}{: 
      year-month-day.
    }
    \item{\code{prec}}{: 
      total precipitation (mm).
    }
  }
}
\details{
  Prior to 1941, precipitation was measured for the 0-24 hour period; from 1941 onwards, precipitation was recorded from 9am to 9am the following day.
}
\references{
  de Carvalho, M., Huser, R., Naveau, P., and Reich, B. J. (2025, to
  appear). \emph{Handbook on Statistics of Extremes}. Chapman &
  Hall/CRC. Boca Raton, FL.
}
\source{
  IPMA (Instituto Português do Mar e da Atmosfera).
}
\keyword{Climatology & Meteorology}
