\encoding{UTF-8}
\name{lse}
\alias{lse}
\title{Selected Stocks from the London Stock Exchange}
\description{
  Prices at close from 26 selected stocks from the London stock exchange
  from 1989 to 2016. 
}
\usage{lse}
\format{
  The \code{lse} data frame has 6894 rows and 27 columns.
}
\references{
  de Carvalho, M., Rubio, R., and Huser (2023).
  Similarity-based clustering for patterns of extreme values.
  Stat, \bold{12}, e560.
}
\keyword{Finance}
